+++ packages/react-native/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @authon/react-native
 
+## 0.8.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @authon/shared@0.8.0
+
 ## 0.7.17
 
 ### Patch Changes